requirements.txt

python-telegram-bot==20.3
openai==1.14.3
langdetect==1.0.9
python-dotenv==1.0.1
